Refactor: import 관련 개선 사항 수정
